% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{Hkangduplicates}
\alias{Hkangduplicates}
\title{Check for Percent Missing}
\usage{
Hkangduplicates(df, variable_list)
}
\arguments{
\item{df}{the analysis dataset being assessed}
}
\value{
A list of duplicates
}
\description{
Assess the duplicates count and list id & visitdate in the analysis dataset.
}
